The great type inquisition - refactoring 99%, adding features 1%.
Please make ambient types work everywhere, more easily, Microsoft.
Ambient types are not being used because I could not manage to get them to work with the frontend, backend, and tests at once.
Most of the time, the simpler solution is better, not the lazier - i.e. ambient types allow me to avoid importing types, for the cost of intense research to get this technological feat to work (months of development).
Also symbolic links don't work (well) with Windows. Mixing Linux and Windows development combined, is not conducive to a working development environment. I know I can use Linux on Windows but that's like using Wine on Linux which is not worth the effort, in my experience.

1) Don't mix Windows and Linux features - they don't work together.
2) Don't use ambient types, unless you need to and understand the full dynamics of ambient types, or are a Microsoft TypeScript engineer.